Changes to project dependencies
Given the complexities with working with jsp in an embedded jetty server, we opted for Tomcat instead. Currently not on the most recent stable version of the server because of some issues around TLD with Tomcat 9. Once we are able to find a solution to that, we will move from version 8 to 9.
And I also created a Runner class to handle the index.jsp file.